Add support for task type aliases
Provide a mapping of aliases to configured types
This will allow users to provide a task via an aliases but still have
the log bundle them under the configured types.

A type can have a many to one relationship
but an alias can/should only be configured to a single type.

Right now it's whichever is processed last is where it's registered to.

Also normalizes each task to lowercase and stripping whitespace: